refactor(tree): default sr strat no circular dep
This changeset removes the circular strategy between the payload
validator and the default state root strategy. Basically we tuck
all the concerns related to spawning the sparse trie task into
the default strategy.
