+++ crates/engine/tree/src/tree/state_root_strategy/sparse_trie.rs
