+++ crates/engine/tree/src/tree/types.rs
@@ -1,9 +1,60 @@
 //! Shared types for blockchain tree validation.
 
 use crate::tree::error::InsertPayloadError;
-use alloy_eip7928::bal::RawBal;
+use alloy_eip7928::bal::{DecodedBal, RawBal};
+use alloy_eips::eip4895::Withdrawal;
+use alloy_primitives::B256;
 use reth_chain_state::{ExecutedBlock, ExecutionTimingStats};
+use reth_evm::{ConfigureEvm, EvmEnvFor};
 use reth_primitives_traits::{BlockTy, NodePrimitives};
+use std::sync::Arc;
+
+/// EVM context required to execute a block.
+#[derive(Debug, Clone)]
+pub struct ExecutionEnv<Evm: ConfigureEvm> {
+    /// Evm environment.
+    pub evm_env: EvmEnvFor<Evm>,
+    /// Hash of the block being executed.
+    pub hash: B256,
+    /// Hash of the parent block.
+    pub parent_hash: B256,
+    /// State root of the parent block.
+    /// Used for sparse trie continuation: if the preserved trie's anchor matches this,
+    /// the trie can be reused directly.
+    pub parent_state_root: B256,
+    /// Number of transactions in the block.
+    /// Used to determine parallel worker count for prewarming.
+    pub transaction_count: usize,
+    /// Total gas used by all transactions in the block.
+    /// Used to adaptively select multiproof chunk size for optimal throughput.
+    pub gas_used: u64,
+    /// Withdrawals included in the block.
+    /// Used to generate prefetch targets for withdrawal addresses.
+    pub withdrawals: Option<Vec<Withdrawal>>,
+    /// Optional decoded BAL for the block.
+    /// Used to validate and optimize execution.
+    pub decoded_bal: Option<Arc<DecodedBal>>,
+}
+
+impl<Evm: ConfigureEvm> ExecutionEnv<Evm>
+where
+    EvmEnvFor<Evm>: Default,
+{
+    /// Creates a new [`ExecutionEnv`] with default values for testing.
+    #[cfg(any(test, feature = "test-utils"))]
+    pub fn test_default() -> Self {
+        Self {
+            evm_env: Default::default(),
+            hash: Default::default(),
+            parent_hash: Default::default(),
+            parent_state_root: Default::default(),
+            transaction_count: 0,
+            gas_used: 0,
+            withdrawals: None,
+            decoded_bal: None,
+        }
+    }
+}
 
 /// Result of block or payload validation.
 pub type ValidationOutcome<N, E = InsertPayloadError<BlockTy<N>>> = Result<ValidationOutput<N>, E>;
